requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dependency-injector==4.46.0
execnet==2.1.1
gherkin-official==29.0.0
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
parse==1.20.2
parse_type==0.6.4
playwright==1.51.0
pluggy==1.5.0
pyee==12.1.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-bdd==8.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.0
pytest-xdist==3.6.1
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
text-unidecode==1.3
typing_extensions==4.13.0
urllib3==2.3.0